Fix issue #1.
Rewrote parts of Egbert Verhage's patches. Thank you again for the work.
This work adds the possibility to generate a INI-style configuration file for the TeamSpeak 3 Server.
The scope is limited to creating and use of the INI config file, and ability to configure network settings, the machine ID and default creation of the first virtual server.
More features can follow soon after.